==> 1_eda_part_1.ipynb <==
import pandas as pd

df_prices_ww = pd.read_parquet('prices_ww.parquet')
df_prices_ww

df_prices_ww['closing_price'].describe()

df_prices_ww['closing_price'].isna().sum()

mask = (df_prices_ww['closing_price'] == 0) | (df_prices_ww['closing_price'].isna())
print(f'удаляем 0 и nan из цен - {mask.sum()} значение ({mask.sum()/len(df_prices_ww) * 100:.2f}%)')
df_prices_ww = df_prices_ww[~mask].copy() # без .copy() pd кидает варнинги при редактировании
df_prices_ww.to_parquet('prices_ww_cleaned.parquet', index=False)
df_prices_ww

# на всякий случай проверям тип
df_prices_ww['closing_price'].apply(lambda x: isinstance(x, float)).all()

df_prices_ww['prev_date'] = df_prices_ww.groupby('stock_id')['date'].shift(1)
df_prices_ww['prev_closing_price'] = df_prices_ww.groupby('stock_id')['closing_price'].shift(1)
df_prices_ww['date_diff'] = (df_prices_ww['date'] - df_prices_ww['prev_date']).dt.days
df_prices_ww['price_change_perc'] = ((df_prices_ww['closing_price']/df_prices_ww['prev_closing_price']) * 100 - 100).round(2)
# -60 and 180 because 1:3 common split, 5 days is max holiday break
df_prices_ww['to_check'] = (df_prices_ww['date_diff'] > 5) | (df_prices_ww['price_change_perc'] < -60) | (df_prices_ww['price_change_perc'] > 180)

to_check = df_prices_ww[df_prices_ww['to_check']]['stock_id'].unique()

print(f'необходимо дополнительно проверить {df_prices_ww['to_check'].sum()} значений для {len(to_check)} акций - их соберем из второго источника данных')

import pickle
with open('to_check.pkl', 'wb') as file:
    pickle.dump(to_check, file)
